Loading more of the premade files
I have been adding only a few files at a time so that I dont
accidentially upload something that is copyrighted by someone else. An
example of this would be the sitemap generator which I had the source
code for but it had a copyright notice. WTF? Anyway I dont want to get
into a legal battle with anyone. Peace!